+++ libs/mops/src/thds/mops/pure/core/deferred_work.py
@@ -9,9 +9,9 @@ from contextlib import contextmanager
 from thds import core
 from thds.core.stack_context import StackContext
 
-_DEFERRED_INVOCATION_WORK: StackContext[ty.Dict[ty.Hashable, ty.Callable[[], ty.Any]]] = StackContext(
-    "DEFERRED_INVOCATION_WORK", dict()
-)
+_DEFERRED_INVOCATION_WORK: StackContext[
+    ty.Optional[ty.Dict[ty.Hashable, ty.Callable[[], ty.Any]]]
+] = StackContext("DEFERRED_INVOCATION_WORK", None)
 logger = core.log.getLogger(__name__)
 
 
@@ -34,12 +34,26 @@ def open_context() -> ty.Iterator[None]:
     existing_work = _DEFERRED_INVOCATION_WORK()
     assert not existing_work, f"deferred work context is not re-entrant! {existing_work}"
     with _DEFERRED_INVOCATION_WORK.set(dict()):
+        logger.debug("Opening deferred work context")
+        yield
+        logger.debug("Closing deferred work context")
+        work_unperformed = _DEFERRED_INVOCATION_WORK()
+        if work_unperformed:
+            logger.warning(
+                f"Internal mops error: some deferred work was not performed before context close! {work_unperformed}"
+            )
+            perform_all()
+
+
+@contextmanager
+def push_non_context() -> ty.Iterator[None]:
+    with _DEFERRED_INVOCATION_WORK.set(None):
         yield
 
 
 def add(work_owner: str, work_id: ty.Hashable, work: ty.Callable[[], ty.Any]) -> None:
     """Add some work to an open context. The work will be performed when perform_all() is
-    called.
+    called. If there is no open context, perform the work immediately.
 
     The work_owner should usually be the module __name__, but if multiple things
     in a module need to add different types of tasks, then it can be anything
@@ -47,14 +61,24 @@ def add(work_owner: str, work_id: ty.Hashable, work: ty.Callable[[], ty.Any]) ->
 
     The work_id should be a unique id within the work_owner 'namespace'.
     """
-    _DEFERRED_INVOCATION_WORK()[(work_owner, work_id)] = work
+    deferred_work = _DEFERRED_INVOCATION_WORK()
+    if deferred_work is None:
+        logger.debug("No open context - performing work %s immediately", work)
+        work()
+    else:
+        logger.debug("Adding work %s to deferred work %s", (work_owner, work_id), id(deferred_work))
+        deferred_work[(work_owner, work_id)] = work
 
 
 def perform_all() -> None:
     """execute all the deferred work that has been added to the current context."""
     work_items = _DEFERRED_INVOCATION_WORK()
     if work_items:
-        logger.info("Performing %s items of deferred work prior to function invocation", len(work_items))
-        for _ in core.parallel.yield_all(work_items.items()):
+        logger.info("Performing %s items of deferred work", len(work_items))
+        for key, _result in core.parallel.yield_all(dict(work_items).items()):
             # consume iterator but don't keep results in memory.
-            pass
+            logger.debug("Popping deferred work %s from %s", key, id(work_items))
+            work_items.pop(key)
+
+        logger.debug("Done performing deferred work on %s", id(work_items))
+        assert not work_items, f"Some deferred work was not performed! {work_items}"