+++ libs/mops/src/thds/mops/pure/core/entry/runner_registry.py
@@ -2,6 +2,7 @@
 
 In practice we only have a single Runner type registered, the MemoizingPicklingRunner.
 """
+
 import typing as ty
 
 from thds.core import stack_context